Rename configs to clientConfig and serverConfig
