+++ client/src/utils/loadData.ts
@@ -9,7 +9,7 @@ import { submitSessionRecovery } from "client/views/login/loginThunks";
 import { store } from "client/utils/store";
 import { AppDispatch } from "client/typings/redux.typings";
 import { submitGetTestSettings } from "client/test/test-settings/testSettingsThunks";
-import { config } from "client/config";
+import { config } from "client/clientConfig";
 
 export const loadData = async (): Promise<void> => {
   // Get app settings
