+++ client/src/views/all-games/components/ExpandedGameDescription.tsx
@@ -5,7 +5,7 @@ import { useAppSelector } from "client/utils/hooks";
 import { FeedbackForm } from "client/views/all-games/components/FeedbackForm";
 import { Game } from "shared/typings/models/game";
 import { UserGroup } from "shared/typings/models/user";
-import { config } from "client/config";
+import { config } from "client/clientConfig";
 
 interface Props {
   game: Game;
