+++ client/webpack.config.ts
@@ -7,7 +7,7 @@ import { Configuration } from "webpack";
 import "webpack-dev-server";
 import { merge } from "webpack-merge";
 import ReactRefreshWebpackPlugin from "@pmmmwh/react-refresh-webpack-plugin";
-import { config } from "./src/config";
+import { config } from "./src/clientConfig";
 
 const TARGET = process.env.npm_lifecycle_event;
