+++ server/src/features/game-popularity/updateGamePopularity.ts
@@ -1,7 +1,7 @@
 import { logger } from "server/utils/logger";
 import { updateWithSignups } from "server/features/game-popularity/utils/updateWithSignups";
 import { updateWithAssign } from "server/features/game-popularity/utils/updateWithAssign";
-import { GameUpdateMethod, config } from "server/config";
+import { GameUpdateMethod, config } from "server/serverConfig";
 import { findUsers } from "server/features/user/userRepository";
 import { findGames } from "server/features/game/gameRepository";
 import { findSignups } from "server/features/signup/signupRepository";
