+++ server/src/features/player-assignment/runAssignment.ts
@@ -6,7 +6,7 @@ import { ProgramType } from "shared/typings/models/game";
 import { findUsers } from "server/features/user/userRepository";
 import { findGames } from "server/features/game/gameRepository";
 import { AssignmentStrategy } from "shared/config/sharedConfig.types";
-import { config } from "server/config";
+import { config } from "server/serverConfig";
 import { removeOverlapSignups } from "server/features/player-assignment/utils/removeOverlapSignups";
 import { saveResults } from "server/features/player-assignment/utils/saveResults";
 import { getDynamicStartTime } from "server/features/player-assignment/utils/getDynamicStartTime";
